zmiany
